+++ R/Model-class.R
@@ -10,18 +10,13 @@ NULL
 #'
 #' @description `r lifecycle::badge("stable")`
 #'
-#' [`AllModels`] is a class from which all the models inherit.
-#'
-#' @slot datanames (`character`)\cr the names of all data slots that are used
-#'   in all the models. In particular, those are also used in the `datamodel` or
-#'   `priormodel` definition for [`GeneralModel`].
+#' [`AllModels`] is a parent class for all models.
 #'
 #' @aliases AllModels
 #' @export
 #'
 .AllModels <- setClass(
-  Class = "AllModels",
-  slots = c(datanames = "character")
+  Class = "AllModels"
 )
 
 # GeneralModel-class ----
@@ -51,6 +46,10 @@ NULL
 #' @slot init (`function`)\cr a function computing the list of starting values
 #'   for parameters required to be initialized in the MCMC sampler, based on the
 #'   data slots that are required as arguments of this function.
+#' @slot datanames (`character`)\cr the names of all data slots that are used
+#'   by `datamodel` JAGS function.
+#' @slot datanames_prior (`character`)\cr the names of all data slots that are
+#'   used by `priormodel` JAGS function.
 #' @slot sample (`character`)\cr names of all parameters from which you would
 #'   like to save the MCMC samples.
 #'
@@ -65,6 +64,8 @@ NULL
     priormodel = "function",
     modelspecs = "function",
     init = "function",
+    datanames = "character",
+    datanames_prior = "character",
     sample = "character"
   ),
   prototype = prototype(
@@ -205,8 +206,8 @@ ModelLogNormal <- function(mean, cov, ref_dose = 1) {
     init = function() {
       list(theta = c(0, 1))
     },
-    sample = c("alpha0", "alpha1"),
-    datanames = c("nObs", "y", "x")
+    datanames = c("nObs", "y", "x"),
+    sample = c("alpha0", "alpha1")
   )
 }
 
@@ -387,8 +388,8 @@ LogisticLogNormalSub <- function(mean, cov, ref_dose = 0) {
     init = function() {
       list(theta = c(0, -20))
     },
-    sample = c("alpha0", "alpha1"),
-    datanames = c("nObs", "y", "x")
+    datanames = c("nObs", "y", "x"),
+    sample = c("alpha0", "alpha1")
   )
 }
 
@@ -589,8 +590,8 @@ LogisticKadane <- function(theta, xmin, xmax) {
     init = function() {
       list(rho0 = theta / 10, gamma = (xmax - xmin) / 2)
     },
-    sample = c("rho0", "gamma"),
-    datanames = c("nObs", "y", "x")
+    datanames = c("nObs", "y", "x"),
+    sample = c("rho0", "gamma")
   )
 }
 
@@ -820,8 +821,8 @@ LogisticNormalMixture <- function(comp1,
     init = function() {
       list(theta = c(0, 1))
     },
-    sample = c("alpha0", "alpha1", "w"),
-    datanames = c("nObs", "y", "x")
+    datanames = c("nObs", "y", "x"),
+    sample = c("alpha0", "alpha1", "w")
   )
 }
 
@@ -963,8 +964,8 @@ LogisticNormalFixedMixture <- function(components,
     init = function() {
       list(theta = c(0, 1))
     },
-    sample = c("alpha0", "alpha1"),
-    datanames = c("nObs", "y", "x")
+    datanames = c("nObs", "y", "x"),
+    sample = c("alpha0", "alpha1")
   )
 }
 
@@ -1068,8 +1069,8 @@ LogisticLogNormalMixture <- function(mean,
     init = function() {
       list(theta = matrix(c(0, 0, 1, 1), nrow = 2))
     },
-    sample = c("alpha0", "alpha1", "comp"),
-    datanames = c("nObs", "y", "x", "nObsshare", "yshare", "xshare")
+    datanames = c("nObs", "y", "x", "nObsshare", "yshare", "xshare"),
+    sample = c("alpha0", "alpha1", "comp")
   )
 }
 
@@ -1264,8 +1265,9 @@ DualEndpoint <- function(mean,
     init = function(y) {
       c(comp$init, list(z = ifelse(y == 0, -1, 1), theta = c(0, 1)))
     },
-    sample = comp$sample,
-    datanames = c("nObs", "w", "x", "xLevel", "y", "nGrid")
+    datanames = c("nObs", "w", "x", "xLevel", "y", "nGrid"),
+    datanames_prior = c("nGrid", "doseGrid"),
+    sample = comp$sample
   )
 }
 
@@ -1962,8 +1964,7 @@ LogisticIndepBeta <- function(binDLE,
     phi1 = phi1,
     phi2 = phi2,
     Pcov = Pcov,
-    data = data,
-    datanames = c("nObs", "y", "x")
+    data = data
   )
 }
 
@@ -2183,8 +2184,7 @@ Effloglog <- function(eff,
     mu = as.vector(mu),
     Q = Q,
     const = const,
-    data = data,
-    datanames = c("nObs", "w", "x")
+    data = data
   )
 }
 
@@ -2353,8 +2353,7 @@ EffFlexi <- function(eff,
     X = X,
     RW = RW,
     RW_rank = RW_rank,
-    data = data,
-    datanames = c("nObs", "w", "x")
+    data = data
   )
 }
 
@@ -2508,8 +2507,8 @@ DALogisticLogNormal <- function(npiece = 3,
     datamodel = datamodel,
     priormodel = priormodel,
     modelspecs = modelspecs,
-    sample = c("alpha0", "alpha1", "lambda"),
-    datanames = c("nObs", "y", "x", "u", "Tmax")
+    datanames = c("nObs", "y", "x", "u", "Tmax"),
+    sample = c("alpha0", "alpha1", "lambda")
   )
 }