+++ NEWS.md
@@ -1,4 +1,6 @@
 # Version 1.0.9000.9015
+* Added new slot `datanames_prior` for `GeneralModel` class.
+* Moved `datanames` slot from `AllModels` class to its child class `GeneralModel`.
 * Renamed several slots in `EffFlexi` and `DALogisticLogNormal` classes and
   adjusted corresponding constructor functions.
 * Renamed `Eff` argument to `eff` and `Effdose` to `eff_dose` in `Effloglog` class